Allow deleting of a component group. Closes #286

+++ app/views/dashboard/incidents/index.blade.php
@@ -29,7 +29,7 @@
                         </div>
                         <div class="col-md-6 text-right">
                             <a href="/dashboard/incidents/{{ $incident->id }}/edit" class="btn btn-default">{{ trans('forms.edit') }}</a>
-                            <a href="/dashboard/incidents/{{ $incident->id }}/delete" class="btn btn-danger">{{ trans('forms.delete') }}</a>
+                            <a href="/dashboard/incidents/{{ $incident->id }}/delete" class="btn btn-danger confirm-action" data-method='DELETE'>{{ trans('forms.delete') }}</a>
                         </div>
                     </div>
                     @endforeach
